Refactor: 레디스 스트림은 프록시 객체를 통해 주입을 받는다.
