+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/service/proxy/MemberRedisProxyService.java
@@ -1,14 +1,11 @@
-package com.sparta.moim.gathering.gathering.application.service.redis;
+package com.sparta.moim.gathering.gathering.application.service.proxy;
 
 import com.sparta.moim.gathering.gathering.application.dto.command.SearchGatheringCommand.JoinGatheringCommand;
 import com.sparta.moim.gathering.gathering.application.dto.command.SearchGatheringCommand.LeaveGatheringCommand;
 import com.sparta.moim.gathering.gathering.application.dto.command.SearchGatheringCommand.RemoveGatheringCommand;
-import com.sparta.moim.gathering.gathering.application.exception.NotOpenGatheringException;
 import com.sparta.moim.gathering.gathering.application.service.MemberService;
+import com.sparta.moim.gathering.gathering.application.service.struct.MemberServiceStruct;
 import com.sparta.moim.gathering.gathering.domain.entity.Member;
-import com.sparta.moim.gathering.gathering.domain.repository.GatheringValidationRepository;
-import com.sparta.moim.gathering.gathering.domain.repository.MemberRepository;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Primary;
@@ -17,46 +14,27 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-//@Primary
-public class MemberServiceAsRedis implements MemberService {
-
+@Primary
+public class MemberRedisProxyService implements MemberService {
   @Value("${spring.data.redis.stream-key}")
   private String streamKey;
-
   private final RedisTemplate<String, Member> redisTemplate;
-  private final MemberRepository memberRepository;
-  private final GatheringValidationRepository gatheringValidationRepository;
+  private final MemberServiceStruct memberServiceStruct;
 
+  @Override
   public void joinGathering(JoinGatheringCommand command) {
-//    validateGatheringExists(command.gatheringId());
-//    statusTrueValidate(command);
-
-//    if (memberRepository.existsByMemberId(command.username())) {
-//      throw new AlreadyParticipateFoundGatheringException();
-//    }
-
     Member member = command.toDomain();
-    // 대기열을 통해 메시지를 전달한다.
+    memberServiceStruct.joinGathering(command);
     redisTemplate.opsForStream().add(streamKey, member.toMap());
   }
 
-  private void statusTrueValidate(JoinGatheringCommand command) {
-    if (!gatheringValidationRepository.isGatheringOpen(command.gatheringId())) {
-      throw new NotOpenGatheringException();
-    }
-  }
-
-  //미구현
+  @Override
   public void leaveGathering(LeaveGatheringCommand command) {
-
+    memberServiceStruct.leaveGathering(command);
   }
 
-  //미구현
+  @Override
   public void removeGathering(RemoveGatheringCommand command) {
-
-  }
-
-  private void validateGatheringExists(UUID id) {
-    gatheringValidationRepository.existsByTrackingIdAndDeletedAtNull(id);
+    memberServiceStruct.removeGathering(command);
   }
 }
