+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/service/struct/InternalMemberService.java
@@ -1,4 +1,4 @@
-package com.sparta.moim.gathering.gathering.application.service.rds;
+package com.sparta.moim.gathering.gathering.application.service.struct;
 
 import com.sparta.moim.gathering.gathering.application.dto.result.GetMemberListResult;
 import com.sparta.moim.gathering.gathering.domain.entity.Member;
